editoast: fga: move parts of client in `api`
