+++ editoast/fga/src/client/api.rs
@@ -1,0 +1,7 @@
+//! Low-level mapping of the OpenFGA HTTP API
+
+pub(super) mod authorization_models;
+pub(super) mod healthz;
+pub(super) mod queries;
+pub(super) mod stores;
+pub(super) mod tuples;
